requirements.txt
Django==2.0.5
django-formtools==2.1
django-admin-env-notice==0.1.3
gunicorn==19.7.1
Pillow==4.2.1
psycopg2==2.7.1
django_compressor==2.1.1
beautifulsoup4==4.6.0

# Tests / linters
django-rainbowtests==0.6.0
pycodestyle==2.4.0
pep8==1.7.0
coverage==4.4.1
coverage-badge==0.2.0
pylama==7.4.1
tblib==1.3.2

raven==6.7.0


# Performance
django-cacheops==3.2.1
django-redis==4.8.0
redis==2.10.5
django-migrations-graph==0.1.1
django-eraserhead==0.1.4
django-debug-toolbar==1.8
